done with the backend functionality

+++ server/controller/Product-controller.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import Product from "../models/Product.js";
 //lets do all the fetching,adding,deleting and updating here
 export const fetchAllProducts = async (req, res) => {
@@ -30,9 +31,32 @@ export const addNewProduct = async (req, res) => {
     return res.status(500).json({ error: "Failed to save product" });
   }
 };
-export const updateProduct = async (req, res) => {};
+export const updateProduct = async (req, res) => {
+  const { id } = req.params;
+  const product = req.body;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ message: "Invalid Product Id" });
+  }
+  let currentProductToUpdate;
+  try {
+    currentProductToUpdate = await Product.findByIdAndUpdate(id, product, {
+      new: true,
+    });
+    res.status(200).json({ data: currentProductToUpdate });
+  } catch (error) {
+    res.status(500).json({ message: "Server Error", error });
+  }
+  if (!currentProductToUpdate) {
+    return res
+      .status(500)
+      .json({ message: "Unable to update. Please try again" });
+  }
+};
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ message: "Invalid Product Id" });
+  }
   try {
     const findCurrentProduct = await Product.findByIdAndDelete(id);
     if (!findCurrentProduct) {
@@ -40,6 +64,6 @@ export const deleteProduct = async (req, res) => {
     }
     return res.status(200).json({ message: "Successfully Deleted" });
   } catch (error) {
-    res.status(404).json({ message: "Unable to delete! Please try again" });
+    res.status(500).json({ message: "Server Error" });
   }
 };
